Update CLV.py
NOTE : The aim of this file is not to support any kind of illegal work , And the developer(Grey_Hat99) is not responsible for any kind of harm done by any person via this tool. This may be punishable offense and anyone can't have the right to sell or promote this product without my permission , if want permission mail me at john.srkpm@gmail.com. 

Available for : Windows only ( Not available for Linux, android or raspberry pi )
Warning : Use at your own risk